requirements.txt
mcp>=1.0.0
openai>=1.0.0
pinecone-client>=3.0.0
pydantic>=2.0.0
uvicorn>=0.20.0
